roots/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anthropic==0.45.2
    # via -r requirements.in
anyio==4.8.0
    # via
    #   anthropic
    #   httpx
    #   openai
bidict==0.23.1
    # via python-socketio
blinker==1.9.0
    # via flask
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
click==8.1.8
    # via flask
distro==1.9.0
    # via
    #   anthropic
    #   openai
faiss-cpu==1.10.0
    # via -r requirements.in
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-socketio
flask-cors==5.0.0
    # via -r requirements.in
flask-socketio==5.5.1
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   wsproto
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   ollama
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
itsdangerous==2.2.0
    # via flask
jinja2==3.1.5
    # via flask
jiter==0.8.2
    # via
    #   anthropic
    #   openai
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
numpy==2.2.3
    # via faiss-cpu
ollama==0.4.7
    # via -r requirements.in
openai==1.63.0
    # via -r requirements.in
packaging==24.2
    # via faiss-cpu
pydantic==2.10.6
    # via
    #   anthropic
    #   ollama
    #   openai
pydantic-core==2.27.2
    # via pydantic
python-engineio==4.11.2
    # via python-socketio
python-socketio==5.12.1
    # via flask-socketio
simple-websocket==1.1.0
    # via python-engineio
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   openai
tenacity==9.0.0
    # via -r requirements.in
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anthropic
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
werkzeug==3.1.3
    # via flask
wsproto==1.2.0
    # via simple-websocket
